refactor service
