+++ src/services/types.ts
@@ -1,0 +1,7 @@
+export interface errorService{
+    error: string;
+    message: string;
+    path: string;
+    status: number;
+    timestamp: string;
+}
